_DevelopmentEnvironment/requirements.txt
# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
alembic>=1.9.0

# API
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.23.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Visualization
plotly>=5.14.0
dash>=2.9.0
streamlit>=1.22.0
pydeck>=0.8.0

# Cache and messaging
redis>=4.5.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.6.0
typer>=0.7.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.1.0
isort>=5.12.0
mypy>=1.2.0
flake8>=6.0.0 